WIP: Open the popup when keybinding fails due to locked vault
Trying to auto-fill or generate a password using Ctrl+Shift+L or
Ctrl+Shift+9, respectively, while the vault is locked should prompt
the user to unlock the vault, instead of silently failing.

Ideally, the popup should close itself after unlocking successfully;
filling in the form in case of the autofill command and copying the
generated password to the clipboard otherwise.

But that would require quite a lot of changes. With this change, at
least there's feedback about the intended action.