fix: Notification bar does not appear after login form change
In some websites (e.g. inoreader.com), the login form is rendered twice,
with the same form id. Thus, only the first form is listened,
while it doesn't exist anymore, so the submit event was missed.
We fix that by storing the forms and forcing a new listening if a new form
with an existing id is different.