fix: Don't override whole jslib file
When building for production, the alias used to override the
sync.service.ts file from the jslib was not resolved. A little research
showed that the problem is caused by the usage of the
AngularCompilerPlugin in the webpack config. But after reading the docs,
I was not able to fix the problem.

So the solution here is to not override the file, but to manually change
the import path in files were it's imported.