feat: Introduce a custom TOTP field type
Allow TOTPs to be entered into custom fields. These are populated based
on an additional field type (TOTP) which will enter the TOTP into the
matched field.

This only takes effect if all of the following are true:

 - There is a TOTP defined for the current cipher
 - The user (or their org) has TOTP functionality
 - A TOTP-based custom field has been defined
 - There are matching field(s) found in the page

The TOTP entry can be customized by the end user through the placement
of a `%s` token. For example `0000%s1111` might generate a TOTP entry of
`00009999991111` if the TOTP was `999999` at the time of form fill.

This change also introduces a simple static HTML file, which can be used
as an autofill playground. By defining a new site in the user's vault
for `file://` URLs, the developer can navigate to
`file://<...>/autofill-playground.html` to easily test the autofill
functionality.