Make autofill doc-scanner traverse into shadowroots
Shadow-roots are a recent invention that makes a web page an
environment that can support independently-operating components.
They give separate namespaces that the well-known element-query
functions work upon, and for which separate searches and answers
are given. It's becoming normal for a web page to create UI struc-
tures that are maintained by separate code, so a search form can
not interfere with a log-in form.

Since searches do not pass from one (document or shadow) root to
other (doc or sh) root, the form username-and-password search
functions need to intentionally decend into each new root as it
finds it.

Closes: #713