Skip build for browser/desktop on importer changes
The importers recently got extracted from libs/common. As the import functionality is currently limited to the web and cli, there is no need to build browser and desktop as there won't be any changes.

This will free up some runners for other workflows (GH actions) to get faster builds that actually impact browser and desktop.

Another benefit is faster feedback on the PR-checks for those two