Allow bitmenu to overflow and scroll when needed
When a user is a member of a lot of organizations and has a small screen
device (or rotated 90 degrees) and want's to select an organization or
maybe even add a new organization, that isn't possible because those
links in the menu are not visible because they overflow the screen.

This PR simply adds the `tw-overflow-y-auto` class to the menu `div` and
will automatically add a scrollbar when needed.