feat: add `step` property to KDF iter input
The copy recommends bumping KDF iterations in increments of 100,000. If
we add `step="100000"` to the HTML, it should cause the little arrows in
the input element to change the value by increments of 100k at a time.

Note: this is a first-time contribution for me, and I do now know how to
test this change locally.